requirements.txt
pandas
openpyxl
python-dotenv
oracledb
apache-airflow